+++ src/Support/CMakeLists.txt
@@ -2,7 +2,7 @@
 
 add_onnx_mlir_library(OMSupport
   KrnlSupport.cpp
-  OptimizeOptions.cpp
+  OMOptions.cpp
 
   LINK_LIBS PUBLIC
   OMKrnlOps
